rename click-handlers.js to article-click-handlers.js
