+++ src/components/SingleArticle/article-click-handlers.js
